Move index.js in src/
